feature/5G-78 redux change data
