+++ packages/client/src/redux/store.ts
@@ -1,11 +1,10 @@
 import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 
-import { appReducer, profileReducer } from './app-reducer'
+import { appReducer } from './app-reducer'
 
 const reducers = combineReducers({
   app: appReducer,
-  profile: profileReducer,
 });
 
 const composeEnchancers =
